Add the first raw code in R programming language
This first code is really crude, but it works with some manually changes. The main goal is to improve this code with functions and automatic rename files names respecting date sequence for 10 days. From D0 to D9.

Example: 

YYYYMMDD_STANDARD_NAME

20190529_standard_name
20190530_standard_name
20190531_standard_name
20190532_standard_name -> This rename is wrong, because doesn't exist "may 32"

How to change automatically to another month from the first day?

-> 20190601_standard_name